pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contextual-rag-chatbot"
version = "0.1.0"
description = "Contextual RAG Chatbot with Open WebUI Integration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "RAG Chatbot Team", email = "team@example.com"}
]
keywords = ["rag", "chatbot", "llm", "vector-search", "document-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.4,<0.3.0",
    "llama-index>=0.9.0",
    "llama-index-vector-stores-postgres>=0.1.0,<0.3.3",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-ollama>=0.1.0",
    "openai>=1.3.0",
    "pymupdf>=1.23.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "crewai>=0.11.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.28",
    "langchain-ollama>=0.1.0",
    "openai>=1.3.0",
    "tiktoken>=0.5.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "arize-phoenix",
    "openinference-instrumentation-openai",
    "docling>=2.0.0",
    "python-pptx>=1.0.2",
    "pymupdf>=1.23.0",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "ragas==0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

# Evaluation dependencies
evaluation = [
    "arize-phoenix>=2.0.0",
    "openinference-instrumentation-openai>=0.1.0",
    # "ragas>=0.0.22",  # Uncomment if needed
]

[project.urls]
Homepage = "https://github.com/your-username/contextual-rag-chatbot"
Repository = "https://github.com/your-username/contextual-rag-chatbot"
Issues = "https://github.com/your-username/contextual-rag-chatbot/issues"

[project.scripts]
rag-chatbot = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg-info",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pymupdf.*",
    "pypdf2.*",
    "docx.*",
    "psycopg2.*",
    "pgvector.*",
]
ignore_missing_imports = true 
